Use line with faded are for densities
The lines will be on top of the areas
so they are always visible even when there are overlapping
distributions.
The shaded area is kept faint
but it is still useful as it puts visual emphasize on the chunk/area
location of the distribution,
instead of just the trend of the line.